requirements.txt
markitdown>=0.1.0a1
langchain>=0.1.0
langchain-ollama>=0.0.3
langchain-core>=0.1.15
psycopg2-binary>=2.9.9
pillow>=10.0.0
markdown>=3.5.1
weasyprint>=60.1
pdf2image>=1.16.3
python-dotenv>=1.0.0